feat(registration): registrations can be finalized (#3)
* feat(registration): registrations can be finalized

* build: align cargo package name with repo name

* build: increase rust version to 2024

* build: replace env::set_var with Justfile

In Rust edition 2024, set_var is considered unsafe. Moving to another solution to set env vars.

* refactor: work around rust issue #100031 with always_send crate

Because of some rust compiler issue, Send cannot be derived correctly for PendingTransactions. Fixing that problem with the use of always_send, a crate that is specifically designed to overcome this kind of problem.

* feat(registration): serialize challenge as base64

* fix(keys): use correct msg when verifying key bundle

* feat(registration): do not wait until account is created on-chain